requirements.txt
Django==3.0.1
django-debug-toolbar==2.1
django-widget-tweaks==1.4.5
django-ipware==2.1.0
mysqlclient==1.4.6
gunicorn==20.0.4
uvicorn==0.11.1